Move BatchType and BatchTypePerseError to scylla-cql-core
BatchType is used extensively in scylla public API.
