+++ scylla-cql-core/src/frame/request/mod.rs
@@ -1,3 +1,4 @@
+pub mod batch;
 pub mod query;
 
 /// Possible requests sent by the client.
